requirements.txt
ultralytics
pillow
openvino
fastapi
uvicorn
numpy
python-multipart
streamlit